add attach and launch unit tests

+++ src/process.cpp
@@ -8,6 +8,19 @@
 #include <filesystem>
 #include <libsdb/process.hpp>
 #include <libsdb/error.hpp>
+#include <libsdb/pipe.hpp>
+#include <cstring>
+
+namespace {
+    /* writes the representation of errno to a pipe with given prefix. Send this message and exit with error code */
+    /* for child process to call and send it through to the pipe */
+    void exit_with_perror(sdb::pipe& channel, std::string const& prefix) {
+        auto message = prefix + ": " + std::strerror(errno); 
+        channel.write(reinterpret_cast<std::byte*> (message.data()), message.size());
+        /* terminate the child process with this exit code */
+        std::exit(-1);
+    }
+}
 
 /* stop reason type methods */
 sdb::stop_reason::stop_reason(int wait_status) 
@@ -45,28 +58,37 @@ sdb::stop_reason sdb::Process::wait_on_signal()
     state_ = reason.reason;
     return reason;
 }
+
 /****************************
 * PROCESS METHODS
 ******************************/
-
 /* launch and attach can call C++ private constructor */
 /* handles fed filepath */
 std::unique_ptr<sdb::Process>
-sdb::Process::launch(std::filesystem::path path) 
+sdb::Process::launch(std::filesystem::path path, bool debug) 
 {
     pid_t pid;
+    /* we need to create pipes before forking. pipes closed when they are not in use */
+    sdb::pipe channel(/*close_on_exec = */true);
+
     if ((pid = fork()) < 0) 
     {
         /* error: fork failed */
         error::send_errno("Process launch failed");
     }
 
+    /******************************/
+    /* CHILD PROCESS */ 
+    /****************************** */
     if (pid == 0) 
-    {
-        if (ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) 
+    {   
+        /* child process won't be using read end only write end */
+        channel.close_read();
+        if (debug and ptrace(PTRACE_TRACEME, 0, nullptr, nullptr) < 0) 
         {
             /* error: tracing failed */
-            error::send_errno("Tracing failed");
+            // error::send_errno("Tracing failed");
+            exit_with_perror(channel, "Tracing failed");
         }
         
         /* gets null-terminated array of characters. C-string */
@@ -74,14 +96,32 @@ sdb::Process::launch(std::filesystem::path path)
         {
             /* error: exec failed */
             /* replaces the current running process with a new process */
-            error::send_errno("Exec failed");
+            // error::send_errno("Exec failed");
+            exit_with_perror(channel, "Exec failed");
         }
     }
 
+    /******************************/
+    /* PARENT PROCESS */ 
+    /****************************** */
+    channel.close_write();
+    auto data = channel.read();
+    channel.close_read();
+
+    /* child process sent an error message, parent process te*/
+    /* for debugging */
+    if (data.size() > 0) {
+        waitpid(pid, nullptr, 0);
+        auto chars = reinterpret_cast<char*>(data.data());
+        error::send(std::string(chars, chars + data.size()));
+    }
     /* terminate_on_end = true as we launched the process */
     //std::unique_ptr<Process> proc = std::make_unique<Process>(pid, /*terminate_on_end=*/true);
-    std::unique_ptr<Process> proc (new Process(pid, true));
-    proc->wait_on_signal();
+    std::unique_ptr<Process> proc (new Process(pid, /*terminate_on_end=*/true, debug));
+
+    if (debug) {
+        proc->wait_on_signal();
+    }
 
     return proc;
 }
@@ -93,6 +133,7 @@ sdb::Process::attach(pid_t pid)
 {
     if (pid == 0) 
     {
+        /* child process */
         /* error: invalid PID */
         error::send("Invalid PID");
     }
@@ -105,7 +146,7 @@ sdb::Process::attach(pid_t pid)
 
     /* terminate_on_end = false as we don't launch the process */
     // std::unique_ptr<Process> proc = std::make_unique<Process>(pid, /*terminate_on_end=*/false);
-    std::unique_ptr<Process> proc (new Process(pid, false));
+    std::unique_ptr<Process> proc (new Process(pid, /*terminate_on_end=*/false, /*attached=*/true));
     proc->wait_on_signal();
 
     return proc;
@@ -128,15 +169,18 @@ sdb::Process::~Process()
     {
         int status;
         /* the process is running with valid PID */
-        if (state_ == process_state::running) 
-        {
+        if (is_attached_) {
+            if (state_ == process_state::running) 
+            {
             /* send it a SIGSTOP and wait for it to stop */
-            kill(pid_, SIGSTOP);
-            waitpid(pid_, &status, 0);
+                kill(pid_, SIGSTOP);
+                waitpid(pid_, &status, 0);
+            }
+            /* detach the destructor (the current process or program) from the process */
+            ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
+            kill(pid_, SIGCONT);
         }
-        /* detach the destructor (the current process or program) from the process */
-        ptrace(PTRACE_DETACH, pid_, nullptr, nullptr);
-        kill(pid_, SIGCONT);
+
 
         /* send it a SIGKILL if we determine that we should destruct it when said program ends*/     
         if (terminate_on_end_)
